src/sitter/seats/serializers.py
from rest_framework import serializers
from .models import User, Seat

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'password']


class SeatSerializer(serializers.ModelSerializer):
    class Meta:
        model = Seat
        fields = ['id', 'name', 'user']

