pyproject.toml
[tool.poetry]
name = "sitter"
version = "0.1.0"
description = ""
authors = ["Bruce McIntyre <bruce.mcintyre@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
django-rest-swagger = "^2.2.0"
qrcode = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
